# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AI Social Media Management Platform API
  version: 0.1.0
  description: Initial REST contract for the enterprise social media management SaaS.
servers:
  - url: http://localhost:4000/api
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /campaigns:
    get:
      tags: [campaigns]
      summary: List workspace campaigns
      parameters:
        - in: query
          name: workspaceId
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Campaign list with operational summaries
  /campaigns/{id}:
    get:
      tags: [campaigns]
      summary: Get campaign details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Campaign details
        "404":
          description: Campaign not found
  /campaigns/{id}/milestones:
    get:
      tags: [campaigns]
      summary: List campaign milestones
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Campaign milestones
  /campaigns/milestones/{milestoneId}/complete:
    post:
      tags: [campaigns]
      summary: Complete a campaign milestone
      parameters:
        - in: path
          name: milestoneId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "201":
          description: Completed milestone
  /campaigns/{id}/tasks:
    get:
      tags: [campaigns]
      summary: List campaign tasks
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Campaign tasks
    post:
      tags: [campaigns]
      summary: Create a campaign task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignTask"
      responses:
        "201":
          description: Created campaign task
  /campaigns/tasks/{taskId}/status:
    post:
      tags: [campaigns]
      summary: Update a campaign task status
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCampaignTaskStatus"
      responses:
        "201":
          description: Updated campaign task
  /campaigns/{id}/budget:
    get:
      tags: [campaigns]
      summary: Campaign budget lines and utilization
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Budget lines, allocated/spent totals, remaining budget, and utilization
  /campaigns/{id}/budget-lines:
    post:
      tags: [campaigns]
      summary: Create or update a campaign budget line
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertCampaignBudgetLine"
      responses:
        "201":
          description: Created or updated budget line
  /campaigns/{id}/reports:
    get:
      tags: [campaigns]
      summary: List generated campaign reports
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Campaign reports
  /campaigns/{id}/reports/generate:
    post:
      tags: [campaigns]
      summary: Generate a campaign report
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateCampaignReport"
      responses:
        "201":
          description: Generated campaign report with metrics and insights
...
  /brand-voices:
    get:
      tags: [brand-voices]
      summary: List workspace brand voice profiles
      parameters:
        - in: query
          name: workspaceId
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Brand voice profiles
    post:
      tags: [brand-voices]
      summary: Create a brand voice profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBrandVoice"
      responses:
        "201":
          description: Created brand voice profile
  /brand-voices/{id}:
    get:
      tags: [brand-voices]
      summary: Get a brand voice profile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Brand voice profile
        "404":
          description: Brand voice not found
    put:
      tags: [brand-voices]
      summary: Update a brand voice profile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBrandVoice"
      responses:
        "200":
          description: Updated brand voice profile
  /brand-voices/{id}/duplicate:
    post:
      tags: [brand-voices]
      summary: Duplicate a brand voice profile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "201":
          description: Duplicated brand voice profile
  /brand-voices/{id}/evaluate:
    post:
      tags: [brand-voices]
      summary: Evaluate copy against a brand voice
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvaluateBrandVoice"
      responses:
        "201":
          description: Brand fit score, matched terms, and recommendations
...
  /listening/summary:
    get:
      tags: [listening]
      summary: Social listening summary
      parameters:
        - in: query
          name: workspaceId
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Monitor counts, mention totals, sentiment mix, reach, and open alerts
  /listening/monitors:
    get:
      tags: [listening]
      summary: List social listening monitors
      parameters:
        - in: query
          name: workspaceId
          schema:
            type: string
            format: uuid
        - in: query
          name: status
          schema:
            type: string
            enum: [active, paused, archived]
      responses:
        "200":
          description: Listening monitor list
    post:
      tags: [listening]
      summary: Create a social listening monitor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateListeningMonitor"
      responses:
        "201":
          description: Created listening monitor
  /listening/monitors/{monitorId}/pause:
    post:
      tags: [listening]
      summary: Pause a social listening monitor
      parameters:
        - in: path
          name: monitorId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "201":
          description: Paused listening monitor
  /listening/monitors/{monitorId}/resume:
    post:
      tags: [listening]
      summary: Resume a social listening monitor
      parameters:
        - in: path
          name: monitorId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "201":
          description: Resumed listening monitor
  /listening/monitors/{monitorId}/archive:
    post:
      tags: [listening]
      summary: Archive a social listening monitor
      parameters:
        - in: path
          name: monitorId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "201":
          description: Archived listening monitor
  /listening/mentions:
    get:
      tags: [listening]
      summary: List detected social mentions
      parameters:
        - in: query
          name: workspaceId
          schema:
            type: string
            format: uuid
        - in: query
          name: monitorId
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Social mention list
    post:
      tags: [listening]
      summary: Ingest a social mention and create an alert when needed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestSocialMention"
      responses:
        "201":
          description: Ingested mention with optional generated alert
  /listening/alerts:
    get:
      tags: [listening]
      summary: List social listening alerts
      parameters:
        - in: query
          name: workspaceId
          schema:
            type: string
            format: uuid
        - in: query
          name: resolved
          schema:
            type: boolean
      responses:
        "200":
          description: Listening alert list
  /listening/alerts/{alertId}/resolve:
    post:
      tags: [listening]
      summary: Resolve a social listening alert
      parameters:
        - in: path
          name: alertId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "201":
          description: Resolved listening alert
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
...
    CreateCampaignTask:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 240
        priority:
          type: string
          enum: [low, normal, high, urgent]
        assigneeId:
          type: string
          format: uuid
        dueDate:
          type: string
          format: date
        metadata:
          type: object
          additionalProperties: true
    UpdateCampaignTaskStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [todo, in_progress, blocked, done]
    UpsertCampaignBudgetLine:
      type: object
      required: [category, allocated, spent]
      properties:
        category:
          type: string
          minLength: 1
          maxLength: 120
        allocated:
          type: number
          minimum: 0
        spent:
          type: number
          minimum: 0
        currency:
          type: string
          example: USD
    GenerateCampaignReport:
      type: object
      properties:
        periodStart:
          type: string
          format: date
        periodEnd:
          type: string
          format: date
...
    CreateListeningMonitor:
      type: object
      required: [workspaceId, type, query]
      properties:
        workspaceId:
          type: string
          format: uuid
        type:
          type: string
          enum: [brand, keyword, hashtag, competitor, influencer]
        query:
          type: string
          minLength: 1
          maxLength: 240
        platforms:
          type: array
          items:
            type: string
            enum: [x, instagram, facebook, linkedin, youtube, tiktok, reddit, pinterest, threads, mastodon, bluesky]
        alertThreshold:
          type: integer
          minimum: 0
          maximum: 100
    IngestSocialMention:
      type: object
      required: [workspaceId, monitorId, platform, author, content, sentiment, reach, engagement]
      properties:
        workspaceId:
          type: string
          format: uuid
        monitorId:
          type: string
          format: uuid
        platform:
          type: string
          enum: [x, instagram, facebook, linkedin, youtube, tiktok, reddit, pinterest, threads, mastodon, bluesky]
        author:
          type: string
          minLength: 1
          maxLength: 180
        content:
          type: string
          minLength: 1
          maxLength: 5000
        url:
          type: string
          format: uri
        sentiment:
          type: string
          enum: [negative, neutral, positive, mixed]
        reach:
          type: integer
          minimum: 0
        engagement:
          type: integer
          minimum: 0
        metadata:
          type: object
          additionalProperties: true
...
    CreateBrandVoice:
      type: object
      required: [workspaceId, name]
      properties:
        workspaceId:
          type: string
          format: uuid
        name:
          type: string
        tone:
          type: object
          additionalProperties: true
        style:
          type: object
          additionalProperties: true
        vocabulary:
          type: object
          properties:
            preferredTerms:
              type: array
              items:
                type: string
            bannedTerms:
              type: array
              items:
                type: string
            industryTerms:
              type: array
              items:
                type: string
        emojiUsage:
          type: string
          enum: [none, light, moderate, expressive]
        ctaPreferences:
          type: object
          additionalProperties: true
        examples:
          type: array
          items:
            type: string
    UpdateBrandVoice:
      type: object
      properties:
        name:
          type: string
        tone:
          type: object
          additionalProperties: true
        style:
          type: object
          additionalProperties: true
        vocabulary:
          type: object
          properties:
            preferredTerms:
              type: array
              items:
                type: string
            bannedTerms:
              type: array
              items:
                type: string
            industryTerms:
              type: array
              items:
                type: string
        emojiUsage:
          type: string
          enum: [none, light, moderate, expressive]
        ctaPreferences:
          type: object
          additionalProperties: true
        examples:
          type: array
          items:
            type: string
    EvaluateBrandVoice:
      type: object
      required: [text]
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 5000
...